+++ modules/claims_api/app/controllers/claims_api/v1/forms/base.rb
@@ -88,23 +88,6 @@ module ClaimsApi
           "#{base_filename}_#{SecureRandom.urlsafe_base64(8)}#{file_extension}"
         end
 
-        def bgs_service
-          bgs = BGS::Services.new(
-            external_uid: target_veteran.participant_id,
-            external_key: target_veteran.participant_id
-          )
-          ClaimsApi::Logger.log('poa', detail: 'bgs-ext service built')
-          bgs
-        end
-
-        def local_bgs_service
-          external_key = target_veteran.participant_id.to_s
-          @local_bgs_service ||= ClaimsApi::LocalBGS.new(
-            external_uid: external_key,
-            external_key:
-          )
-        end
-
         def received_date
           form_attributes['received_date']
         end
